Small distrib changes,   Use GTK's native insert for wxListBox,   Compile pacthes for AIX,   wxClipboard things,   Ignore wxSIZE_AUTO flags in wxDialog and wxFrame,   Ignore toolbar size in wxFrame::Get/SetClientSize
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@4162 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/unix/dialup.cpp
@@ -124,7 +124,7 @@ public:
    // NB: this won't result in DISCONNECTED event being sent
    virtual bool CancelDialing();
 
-   unsigned int GetISPNames(class wxArrayString &) const
+   size_t GetISPNames(class wxArrayString &) const
       { return 0; }
    
    // sometimes the built-in logic for determining the online status may fail,
